// frontend/lib/api/generated/model/contratoDetailOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { ContratoDetailOutPenalidadJson } from './contratoDetailOutPenalidadJson';
import type { CronogramaEntregaOut } from './cronogramaEntregaOut';
import type { ProductoContratadoOut } from './productoContratadoOut';

export interface ContratoDetailOut {
  contrato_id: number;
  numero_contrato: string;
  proveedor_id: number;
  requerimiento_anual_id: number;
  fecha_inicio: string;
  fecha_fin: string;
  presupuesto_total: number;
  condiciones: string | null;
  penalidad_json: ContratoDetailOutPenalidadJson;
  responsable_id: number;
  estado: string;
  creado_en: string;
  dias_para_vencer: number;
  alerta_vigencia: boolean;
  proveedor_razon_social: string;
  proveedor_ruc: string;
  cronograma: CronogramaEntregaOut[];
  productos_contratados: ProductoContratadoOut[];
}